requirements.txt
numpy
pandas
PyYAML
scikit-learn
tqdm
torch
transformers
datasets
flask
flask-cors
joblib
pyarrow
dvc
matplotlib
seaborn
plotly
python-json-logger
sentence-transformers
faiss-cpu
# mlflow<3
